[MOB-880] Created Service Builder and moved all services there
